requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
idna==3.10
mysql-connector-python==8.4.0
pycparser==2.22
python-snap7==2.0.2
requests==2.32.3
urllib3==2.3.0
